added create cost dialog
