+++ src/components/createCost/CreateCost.module.css
@@ -2,6 +2,7 @@
     display: flex;
     flex-direction: column;
     justify-content: center;
+    padding: 2rem;
 }
 
 .row {
