+++ src/components/scaffold/drawerContent.js
@@ -1,5 +1,4 @@
 import { trans } from "../../trans/trans";
-import ReceiptIcon from '@material-ui/icons/Receipt';
 import MonetizationOnIcon from '@material-ui/icons/MonetizationOn';
 import paths from "../../routes/paths";
 
@@ -7,13 +6,8 @@ export default [
     [
         {
             title: trans('Components.drawer.listPrices'),
-            icon: ReceiptIcon,
+            icon: MonetizationOnIcon,
             to: paths.costList
         },
-        {
-            title: trans('Components.drawer.createPrice'),
-            icon: MonetizationOnIcon,
-            to: paths.createCost
-        }
     ]
 ];
