+++ src/static/currencies.js
@@ -1,10 +1,10 @@
 export const currencies = [
     {
-        value: 0,
+        value: '0',
         label: 'Bs',
     },
     {
-        value: 1,
+        value: '1',
         label: '$',
     },
 ];
