+++ src/utils/validator/validatorRules.js
@@ -1,7 +1,6 @@
 import { hasContent } from "../common";
 
 export function required(value) {
-    console.log(hasContent(value));
     return hasContent(value);
 }
